requirements.txt
fastapi[all]                    # Включает fastapi, pydantic, uvicorn и другие зависимости
psycopg2[binary]                 # Драйвер для PostgreSQL (можно использовать psycopg2-binary)
redis                           # Для возможного кэширования
alembic                         # Миграции базы данных
passlib[argon2]                 # Хеширование паролей с Argon2 (bcrypt не нужен)
python-jose[cryptography]       # Для работы с JWT-токенами
python-dotenv                   # Чтение конфигурации из .env
pyjwt                           # Генерация JWT-токенов